Se saca interfaz cmsis, se cambian tipos de datos y se modifica stack de tareas

+++ freeRTOSHuerta/Core/Inc/DHT.h
@@ -12,8 +12,8 @@
 
 typedef struct
 {
-	uint32_t Temperature;
-	uint32_t Humidity;
+	uint8_t Temperature;
+	uint8_t Humidity;
 }DHT_DataTypeDef;
